+++ src/main/java/ru/stepanov/skypro/hwcalculator1906/service/CalculatorService.java
@@ -1,8 +1,8 @@
 package ru.stepanov.skypro.hwcalculator1906.service;
 
 public interface CalculatorService {
-    String plus(Integer num1, Integer num2);
-    String minus(Integer num1, Integer num2);
-    String divide(Integer num1, Integer num2);
-    String multiply(Integer num1, Integer num2);
+    String plus(Double num1, Double num2);
+    String minus(Double num1, Double num2);
+    String divide(Double num1, Double num2);
+    String multiply(Double num1, Double num2);
 }
